SortedMultiSet: better return value from Count()

+++ stingraykit/collection/IMultiSet.h
@@ -33,7 +33,7 @@ namespace stingray
 		virtual ~IReadonlyMultiSet() { }
 
 		virtual bool Contains(const ValueType& value) const = 0;
-		virtual int Count(const ValueType& value) const = 0;
+		virtual size_t Count(const ValueType& value) const = 0;
 
 		virtual shared_ptr<IEnumerator<T> > Find(const T& value) const = 0;
 		virtual shared_ptr<IEnumerator<T> > ReverseFind(const T& value) const = 0;
